Convert description metadata in pipe
